import Data-Random 0.04 from CPAN
git-cpan-module:   Data-Random
git-cpan-version:  0.04
git-cpan-authorid: ADEO
git-cpan-file:     authors/id/A/AD/ADEO/Data-Random-0.04.tar.gz

+++ t/rand_image.t
@@ -7,24 +7,26 @@ use lib qw(..);
 use Data::Random qw( rand_image );
 use File::Spec;
 
-print $Data::Random::VERSION,"\n";
-
 use vars qw( $imagefile );
 
-$imagefile = File::Spec->tmpdir().'/Data_Random_'.time().'.tmp';
+# Try to load GD
+eval q{ use GD };
 
+# If the module cannot be loaded, skip tests
+print "1..0 # Skipped: GD not installed\n" and exit if $@;
+
+$imagefile = File::Spec->tmpdir() . '/Data_Random_' . time() . '.tmp';
 
 # Test writing an image to a file
 {
-    open(FILE, ">$imagefile");
+    open( FILE, ">$imagefile" );
     binmode(FILE);
-    print FILE rand_image( bgcolor => [0, 0, 0] );
+    print FILE rand_image( bgcolor => [ 0, 0, 0 ] );
     close(FILE);
-    
-    ok(!(-z $imagefile));
+
+    ok( !( -z $imagefile ) );
 }
 
-
 END {
-    unlink($imagefile);
+    unlink($imagefile) if $imagefile;
 }
